requirements.txt
gradio==5.23.3
google-genai==1.9.0
pathlib==1.0.1